lowering: split off RustIrError to another file
